requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
tqdm>=4.65.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
kaggle>=1.5.12
jupytext>=0.12.0
scikit-image>=0.20.0
